fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-dotenv
pydantic
alembic
bcrypt
python-jose[cryptography]
boto3
python-multipart
slowapi